refactor: remove Python bridge files and subprocess fallback (#2181)
Remove dead Python bridge code now that native Rust transports are
operational (PR #2172). This completes the bridge migration tracked
by epic #2155.

Deleted files:
- python/bridge_server.py, simard_gym_bridge.py, simard_knowledge_bridge.py
- tests/test_bridge_server.py, test_bridges.py, test_knowledge_bridge.py

Rust changes:
- bridge_launcher.rs: removed find_python_dir(), build_python_path(),
  set_python_path(), make_transport(), Python subprocess launchers,
  and Python fallback logic from launch_all_bridges()
- operator_commands_ooda/daemon: use launch_*_native() directly
- ooda_loop/bridge_factory: use launch_*_native() directly
- Updated doc comments in gym_bridge.rs, knowledge_bridge.rs,
  native_gym.rs to remove Python references
- Removed hardcoded /home/azureuser paths (build_python_path deleted)

Infrastructure:
- scripts/check-rust-only-gate.sh: cleared Python allow-list
- tests/fixtures/echo_bridge.py: minimal echo server for
  SubprocessBridgeTransport integration tests
- Updated QA scenario and architecture docs
