+++ docs/architecture/bridge-pattern.md
@@ -1,29 +1,24 @@
 ---
 title: Bridge Pattern
-description: How Simard communicates with the Python ecosystem through subprocess bridges with JSON-line protocol and circuit breaker fault tolerance.
-last_updated: 2026-04-13
+description: How Simard's bridge infrastructure provides typed interfaces for knowledge and gym services, using native Rust transports with circuit breaker fault tolerance.
+last_updated: 2026-06-02
 owner: simard
 doc_type: concept
 ---
 
 # Bridge Pattern
 
-Simard is Rust. The amplihack ecosystem (memory-lib, kg-packs, agent-eval) is Python. Rather than rewrite the ecosystem in Rust or use FFI, Simard communicates through **subprocess bridges** — Python processes that speak a simple JSON-line protocol on stdin/stdout.
+Simard uses a **bridge abstraction** — typed `BridgeTransport` implementations that speak a JSON-line protocol — to isolate bridge-client code from transport details.
 
-## Why Bridges?
+## Transport Types
 
-| Approach | Pros | Cons |
-|----------|------|------|
-| **PyO3 FFI** | Zero-copy, native speed | Tight coupling, GIL contention, complex build |
-| **HTTP/gRPC** | Standard, debuggable | Server lifecycle, port management, overhead |
-| **Subprocess bridges** | Simple, isolated, no dependencies | Serialization overhead, process lifecycle |
+| Transport | Use Case | Notes |
+|-----------|----------|-------|
+| **NativeBridgeTransport** | Production (knowledge, gym) | In-process Rust handlers, zero overhead |
+| **SubprocessBridgeTransport** | Testing infrastructure | Spawns a Python subprocess; used only in integration tests |
+| **InMemoryBridgeTransport** | Unit testing | In-memory mock; no I/O |
 
-Bridges win because:
-- The Python ecosystem already works — we don't want to port 3,000+ LOC of production Python code
-- Process isolation means a Python crash can't take down Simard
-- The circuit breaker pattern handles intermittent failures gracefully
-
-> **Note**: The memory bridge has been replaced by a native Rust implementation (`NativeCognitiveMemory`) that talks directly to LadybugDB via the `lbug` crate. See [Cognitive Memory Architecture](cognitive-memory.md). The bridge pattern remains in use for knowledge and gym bridges.
+> **History**: Prior to #2181, knowledge and gym bridges used Python subprocess transports with a native Rust fallback. The native Rust transports are now the only production path. Memory was already migrated to native `NativeCognitiveMemory` (see [Cognitive Memory Architecture](cognitive-memory.md)).
 
 ## Wire Protocol
 
@@ -158,12 +153,12 @@ let response = transport.call(health_request()).unwrap();
 assert!(response.result.is_some());
 ```
 
-### Integration Tests (real Python subprocess)
+### Integration Tests (subprocess transport)
 
 ```rust
 let transport = SubprocessBridgeTransport::new(
     "echo-test",
-    "python/bridge_server.py",
+    "tests/fixtures/echo_bridge.py",
     vec![],
     Duration::from_secs(5),
 );
