+++ scripts/check-rust-only-gate.sh
@@ -15,10 +15,10 @@ set -euo pipefail
 # ── Allow-list: pre-existing files that are permitted until migrated ──────────
 # Each entry is tracked by the Rust-only epic (#2155). Remove entries as the
 # corresponding rewrite issues (#2156, #2157) are completed.
+#
+# Python bridge files (bridge_server.py, simard_gym_bridge.py,
+# simard_knowledge_bridge.py) removed in #2181 — native Rust transport complete.
 ALLOWED_PY_FILES=(
-  "python/bridge_server.py"
-  "python/simard_gym_bridge.py"
-  "python/simard_knowledge_bridge.py"
 )
 
 ALLOWED_JS_TS_FILES=(
