+++ src/bridge_launcher.rs
@@ -1,97 +1,26 @@
 //! Launch and manage bridge transports for live Simard operations.
 //!
 //! This module provides functions to create bridge transport instances for the
-//! knowledge and gym bridges. The default strategy is to use
-//! [`NativeBridgeTransport`] for in-process Rust execution, falling back to
-//! [`SubprocessBridgeTransport`] (Python) if the native transport fails its
-//! health check.
+//! knowledge and gym bridges using [`NativeBridgeTransport`] for in-process
+//! Rust execution.
 //!
-//! Cognitive memory is now handled natively by [`NativeCognitiveMemory`](crate::cognitive_memory::NativeCognitiveMemory).
+//! Cognitive memory is handled natively by [`NativeCognitiveMemory`](crate::cognitive_memory::NativeCognitiveMemory).
 
-use std::path::{Path, PathBuf};
-use std::sync::Once;
+use std::path::PathBuf;
 use std::time::Duration;
 
 use crate::bridge::BridgeTransport;
 use crate::bridge_circuit::{CircuitBreakerConfig, CircuitBreakerTransport};
-use crate::bridge_subprocess::{NativeBridgeTransport, SubprocessBridgeTransport};
+use crate::bridge_subprocess::NativeBridgeTransport;
 use crate::error::SimardResult;
 use crate::gym_bridge::GymBridge;
 use crate::knowledge_bridge::KnowledgeBridge;
-
-const DEFAULT_BRIDGE_TIMEOUT: Duration = Duration::from_secs(30);
-/// Gym bridge runs full progressive benchmark suites which call out to LLMs;
-/// each suite typically takes several minutes. The 30s default would always
-/// trip and silently disable benchmark feedback.
-const GYM_BRIDGE_TIMEOUT: Duration = Duration::from_secs(900);
 
 fn default_circuit_breaker() -> CircuitBreakerConfig {
     CircuitBreakerConfig {
         failure_threshold: 3,
         cooldown: Duration::from_secs(30),
     }
-}
-
-/// Locate the `python/` directory by walking up from the working directory.
-pub fn find_python_dir() -> SimardResult<PathBuf> {
-    let cwd = std::env::current_dir().unwrap_or_else(|_| PathBuf::from("."));
-    for ancestor in cwd.ancestors() {
-        let candidate = ancestor.join("python").join("bridge_server.py");
-        if candidate.exists() {
-            return Ok(ancestor.join("python"));
-        }
-    }
-    Err(crate::error::SimardError::BridgeSpawnFailed {
-        bridge: "launcher".to_string(),
-        reason: "could not find python/ directory with bridge_server.py".to_string(),
-    })
-}
-
-/// Build PYTHONPATH with ecosystem dependencies.
-fn build_python_path() -> String {
-    let candidates = [
-        "/home/azureuser/src/amplirusty/amplihack-memory-lib/src",
-        "/home/azureuser/src/agent-kgpacks",
-        "/home/azureuser/src/amplihack/src",
-        "/home/azureuser/.amplihack/src",
-    ];
-    let mut paths: Vec<String> = candidates
-        .iter()
-        .filter(|p| Path::new(p).exists())
-        .map(|p| p.to_string())
-        .collect();
-    if let Ok(existing) = std::env::var("PYTHONPATH") {
-        paths.push(existing);
-    }
-    paths.join(":")
-}
-
-fn set_python_path() {
-    static INIT: Once = Once::new();
-    INIT.call_once(|| {
-        // SAFETY: called exactly once during single-threaded bootstrap,
-        // before any bridge subprocesses are spawned.
-        unsafe {
-            std::env::set_var("PYTHONPATH", build_python_path());
-        }
-    });
-}
-
-fn make_transport(name: &str, script: &Path, extra_args: Vec<String>) -> Box<dyn BridgeTransport> {
-    make_transport_with_timeout(name, script, extra_args, DEFAULT_BRIDGE_TIMEOUT)
-}
-
-fn make_transport_with_timeout(
-    name: &str,
-    script: &Path,
-    extra_args: Vec<String>,
-    timeout: Duration,
-) -> Box<dyn BridgeTransport> {
-    let transport = SubprocessBridgeTransport::new(name, script, extra_args, timeout);
-    Box::new(CircuitBreakerTransport::new(
-        transport,
-        default_circuit_breaker(),
-    ))
 }
 
 /// Wrap a native transport in a circuit breaker.
@@ -130,37 +59,19 @@ fn resolve_packs_dir() -> PathBuf {
 /// Launch all bridges, returning None for any that fail (honest degradation).
 ///
 /// Cognitive memory is now native — only knowledge and gym bridges are launched.
-/// The default strategy is native Rust transport, with Python subprocess fallback.
+/// Both use the native Rust transport exclusively.
 pub fn launch_all_bridges(
     _agent_name: &str,
-    _state_root: &Path,
+    _state_root: &std::path::Path,
 ) -> (Option<KnowledgeBridge>, Option<GymBridge>) {
     let knowledge = match launch_knowledge_bridge_native() {
         Ok(b) => {
             eprintln!("[simard] knowledge bridge: using native Rust transport");
             Some(b)
         }
-        Err(native_err) => {
-            eprintln!("[simard] knowledge bridge native transport failed: {native_err}");
-            // Fall back to Python subprocess
-            match find_python_dir() {
-                Ok(python_dir) => match launch_knowledge_bridge_subprocess(&python_dir) {
-                    Ok(b) => {
-                        eprintln!("[simard] knowledge bridge: fell back to Python subprocess");
-                        Some(b)
-                    }
-                    Err(e) => {
-                        eprintln!(
-                            "[simard] knowledge bridge launch FAILED — domain knowledge disabled: {e}"
-                        );
-                        None
-                    }
-                },
-                Err(e) => {
-                    eprintln!("[simard] knowledge bridge launch FAILED — no Python dir: {e}");
-                    None
-                }
-            }
+        Err(e) => {
+            eprintln!("[simard] knowledge bridge launch FAILED — domain knowledge disabled: {e}");
+            None
         }
     };
 
@@ -169,25 +80,9 @@ pub fn launch_all_bridges(
             eprintln!("[simard] gym bridge: using native Rust transport");
             Some(b)
         }
-        Err(native_err) => {
-            eprintln!("[simard] gym bridge native transport failed: {native_err}");
-            // Fall back to Python subprocess
-            match find_python_dir() {
-                Ok(python_dir) => match launch_gym_bridge_subprocess(&python_dir) {
-                    Ok(b) => {
-                        eprintln!("[simard] gym bridge: fell back to Python subprocess");
-                        Some(b)
-                    }
-                    Err(e) => {
-                        eprintln!("[simard] gym bridge launch FAILED — benchmarks disabled: {e}");
-                        None
-                    }
-                },
-                Err(e) => {
-                    eprintln!("[simard] gym bridge launch FAILED — no Python dir: {e}");
-                    None
-                }
-            }
+        Err(e) => {
+            eprintln!("[simard] gym bridge launch FAILED — benchmarks disabled: {e}");
+            None
         }
     };
 
@@ -223,68 +118,9 @@ pub fn launch_gym_bridge_native() -> SimardResult<GymBridge> {
     Ok(GymBridge::new(wrapped))
 }
 
-/// Launch a knowledge graph pack bridge via Python subprocess (fallback).
-pub fn launch_knowledge_bridge(python_dir: &Path) -> SimardResult<KnowledgeBridge> {
-    launch_knowledge_bridge_subprocess(python_dir)
-}
-
-fn launch_knowledge_bridge_subprocess(python_dir: &Path) -> SimardResult<KnowledgeBridge> {
-    set_python_path();
-    let script = python_dir.join("simard_knowledge_bridge.py");
-    let transport = make_transport("knowledge", &script, vec![]);
-    if !check_health("knowledge", transport.as_ref()) {
-        return Err(crate::error::SimardError::BridgeSpawnFailed {
-            bridge: "knowledge".to_string(),
-            reason: "bridge unhealthy after launch".to_string(),
-        });
-    }
-    Ok(KnowledgeBridge::new(transport))
-}
-
-/// Launch a gym/eval bridge via Python subprocess (fallback).
-pub fn launch_gym_bridge(python_dir: &Path) -> SimardResult<GymBridge> {
-    launch_gym_bridge_subprocess(python_dir)
-}
-
-fn launch_gym_bridge_subprocess(python_dir: &Path) -> SimardResult<GymBridge> {
-    set_python_path();
-    let script = python_dir.join("simard_gym_bridge.py");
-    let transport = make_transport_with_timeout("gym-eval", &script, vec![], GYM_BRIDGE_TIMEOUT);
-    if !check_health("gym", transport.as_ref()) {
-        return Err(crate::error::SimardError::BridgeSpawnFailed {
-            bridge: "gym-eval".to_string(),
-            reason: "bridge unhealthy after launch".to_string(),
-        });
-    }
-    Ok(GymBridge::new(transport))
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn find_python_dir_from_repo_root() {
-        let result = find_python_dir();
-        assert!(result.is_ok(), "should find python/ from repo root");
-        assert!(result.unwrap().join("bridge_server.py").exists());
-    }
-
-    #[test]
-    fn build_python_path_returns_string() {
-        let path = build_python_path();
-        let _ = path;
-    }
-
-    #[test]
-    fn default_bridge_timeout_is_30_seconds() {
-        assert_eq!(DEFAULT_BRIDGE_TIMEOUT, Duration::from_secs(30));
-    }
-
-    #[test]
-    fn gym_bridge_timeout_allows_full_suite() {
-        assert!(GYM_BRIDGE_TIMEOUT >= Duration::from_secs(300));
-    }
 
     #[test]
     fn default_circuit_breaker_has_expected_threshold() {
@@ -298,48 +134,10 @@ mod tests {
         assert_eq!(config.cooldown, Duration::from_secs(30));
     }
 
-    #[test]
-    fn build_python_path_is_colon_separated() {
-        let path = build_python_path();
-        if !path.is_empty() {
-            for segment in path.split(':') {
-                assert!(!segment.is_empty(), "path segment should not be empty");
-            }
-        }
-    }
-
-    #[test]
-    fn build_python_path_includes_existing_pythonpath() {
-        let original = std::env::var("PYTHONPATH").ok();
-        unsafe {
-            std::env::set_var("PYTHONPATH", "/test/custom/path");
-        }
-        let path = build_python_path();
-        assert!(path.contains("/test/custom/path"));
-        match original {
-            Some(val) => unsafe {
-                std::env::set_var("PYTHONPATH", val);
-            },
-            None => unsafe {
-                std::env::remove_var("PYTHONPATH");
-            },
-        }
-    }
-
-    #[test]
-    fn find_python_dir_returns_directory_containing_bridge_server() {
-        if let Ok(dir) = find_python_dir() {
-            assert!(dir.is_dir());
-            assert!(dir.join("bridge_server.py").exists());
-        }
-    }
-
     // ── Native transport tests ──
 
     #[test]
     fn launch_knowledge_bridge_native_succeeds() {
-        // Native knowledge bridge should always pass health check
-        // since it registers a bridge.health handler.
         let result = launch_knowledge_bridge_native();
         assert!(
             result.is_ok(),
