+++ tests/qa-scenarios/knowledge-bridge-health.yaml
@@ -1,9 +1,9 @@
 name: simard-knowledge-bridge-health
 app_type: cli
-description: Verify Python knowledge bridge starts and responds to health
+description: Verify native knowledge bridge starts and responds to health
 steps:
   - action: run
-    command: "echo '{\"id\":\"1\",\"method\":\"bridge.health\",\"params\":{}}' | python3 python/simard_knowledge_bridge.py"
-    timeout: 15s
+    command: "cargo test -p simard launch_knowledge_bridge_native_succeeds --no-fail-fast -- --nocapture 2>&1 | tail -5"
+    timeout: 120s
   - action: expect_stdout
-    contains: '"healthy":true'
+    contains: 'ok'
